safari-534.53-branch/LayoutTests/java/lc3/JavaObject/JavaObjectFieldOrMethod-001-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject Field or method access

FAIL dt.amIAFieldOrAMethod should be FIELD! (of type string). Was FIELD! (of type object).
FAIL dt.amIAFieldOrAMethod() should be METHOD!. Threw exception TypeError: 'FIELD!' is not a function (evaluating 'dt.amIAFieldOrAMethod()')
PASS successfullyParsed is true

TEST COMPLETE
 
